[TOK-14] Add redirection to Google provider
