--- src/app/[locale]/components/SignInWithGoogle.tsx
@@ -1,9 +1,0 @@
-"use client";
-
-import { signIn } from "next-auth/react";
-
-const SignInWithGoogle = () => {
-  return <>{signIn("google")}</>;
-};
-
-export default SignInWithGoogle;
