+++ src/app/[locale]/page.tsx
@@ -1,5 +1,9 @@
-const Home = () => {
-  return <h2>MAIN PAGE</h2>;
+import { getSession } from "@/lib/authentication/session";
+
+const Home = async () => {
+  const session = await getSession();
+
+  return <h2>WELCOME {session?.user?.first_name}</h2>;
 };
 
 export default Home;
